iox-#1032 Port 'hoofs' to the new error reporting
