+++ iceoryx_binding_c/test/moduletests/test_chunk.cpp
@@ -20,7 +20,7 @@ extern "C" {
 #include "iceoryx_binding_c/runtime.h"
 }
 
-#include "iceoryx_hoofs/error_handling/error_handling.hpp"
+#include "iox/detail/hoofs_error_reporting.hpp"
 #include "iceoryx_hoofs/testing/fatal_failure.hpp"
 #include "iceoryx_posh/roudi_env/minimal_roudi_config.hpp"
 #include "iceoryx_posh/testing/roudi_gtest.hpp"
