+++ iceoryx_binding_c/test/moduletests/test_request_header.cpp
@@ -14,7 +14,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include "iceoryx_hoofs/error_handling/error_handling.hpp"
+#include "iox/detail/hoofs_error_reporting.hpp"
 #include "iceoryx_hoofs/testing/fatal_failure.hpp"
 #include "iceoryx_posh/popo/rpc_header.hpp"
 #include "iceoryx_posh/testing/mocks/chunk_mock.hpp"
