+++ iceoryx_binding_c/test/moduletests/test_runtime.cpp
@@ -18,7 +18,7 @@ extern "C" {
 #include "iceoryx_binding_c/runtime.h"
 }
 
-#include "iceoryx_hoofs/error_handling/error_handling.hpp"
+#include "iox/detail/hoofs_error_reporting.hpp"
 #include "iceoryx_hoofs/testing/fatal_failure.hpp"
 #include "iceoryx_posh/iceoryx_posh_types.hpp"
 #include "iceoryx_posh/roudi_env/minimal_roudi_config.hpp"
