+++ iceoryx_hoofs/reporting/include/iox/detail/hoofs_error_reporting.hpp
@@ -1,0 +1,106 @@
+// Copyright (c) 2019 - 2020 by Robert Bosch GmbH. All rights reserved.
+// Copyright (c) 2020 - 2022 by Apex.AI Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// SPDX-License-Identifier: Apache-2.0
+#ifndef IOX_HOOFS_ERROR_HANDLING_ERROR_HANDLING_HPP
+#define IOX_HOOFS_ERROR_HANDLING_ERROR_HANDLING_HPP
+
+// Each module (= some unit with its own errors) must provide the following.
+
+// 1. Define the errors of the module -> see below
+
+// 2. Include the custom reporting implementation
+#include "iox/error_reporting/custom/error_reporting.hpp"
+
+// 3. Include the error reporting macro API
+#include "iox/error_reporting/macros.hpp"
+
+// additional includes
+#include "iox/error_reporting/types.hpp"
+
+/// @todo iox-#1032 Remove once HOOFS_MODULE_IDENTIFIER is moved to 'ModuleId' in 'error_reporting/types.hpp'
+#include "iceoryx_hoofs/error_handling/error_handler.hpp"
+
+namespace iox
+{
+// clang-format off
+#define HOOFS_ERRORS(error) \
+    error(EXPECTS_ENSURES_FAILED) \
+    error(DO_NOT_USE_AS_ERROR_THIS_IS_AN_INTERNAL_MARKER) // keep this always at the end of the error list
+
+    // EXPECTS_ENSURES_FAILED is used as a temporary solution to make IOX_EXPECTS/IOX_ENSURES testable
+
+// clang-format on
+
+// DO NOT TOUCH THE ENUM, you can doodle around with the lines above!!!
+
+enum class HoofsError : iox::er::ErrorCode::type
+{
+    HOOFS_ERRORS(CREATE_ICEORYX_ERROR_ENUM)
+};
+
+const char* asStringLiteral(const HoofsError error) noexcept;
+
+class HoofsErrorType
+{
+  public:
+    explicit HoofsErrorType(HoofsError code)
+        : m_code(static_cast<iox::er::ErrorCode::type>(code))
+    {
+    }
+
+    static constexpr iox::er::ModuleId module()
+    {
+        return MODULE_ID;
+    }
+
+    iox::er::ErrorCode code() const
+    {
+        return m_code;
+    }
+
+    const char* name() const
+    {
+        return asStringLiteral(static_cast<HoofsError>(m_code.value));
+    }
+
+    const char* moduleName() const {
+        return "iceoryx_hoofs";
+    }
+
+    static constexpr iox::er::ModuleId MODULE_ID{POSH_MODULE_IDENTIFIER};
+
+  protected:
+    iox::er::ErrorCode m_code;
+};
+
+namespace er
+{
+
+inline HoofsErrorType toError(HoofsError code)
+{
+    return HoofsErrorType(code);
+}
+
+inline ModuleId toModule(HoofsError)
+{
+    return HoofsErrorType::MODULE_ID;
+}
+
+} // namespace er
+
+} // namespace iox
+
+#endif // IOX_HOOFS_ERROR_HANDLING_ERROR_HANDLING_HPP
