+++ iceoryx_hoofs/test/moduletests/test_container_fixed_position_container.cpp
@@ -17,7 +17,7 @@
 
 #include "iox/fixed_position_container.hpp"
 
-#include "iceoryx_hoofs/error_handling/error_handling.hpp"
+#include "iox/detail/hoofs_error_reporting.hpp"
 #include "iceoryx_hoofs/testing/fatal_failure.hpp"
 #include "iceoryx_hoofs/testing/lifetime_and_assignment_tracker.hpp"
